+++ src/app/app-routing.module.ts
@@ -19,6 +19,7 @@ import { adminGuard } from "./auth/guards/admin.guard";
 import { AdminDashboardComponent } from "./admin-view/admin-dashboard/admin-dashboard.component";
 import { DatasetFlowDetailsComponent } from "./dataset-flow/dataset-flow-details/dataset-flow-details.component";
 import { AccountComponent } from "./account/account.component";
+import { QueryExplainerComponent } from "./query-explainer/query-explainer.component";
 
 export const routes: Routes = [
     { path: "", redirectTo: ProjectLinks.DEFAULT_URL, pathMatch: "full" },
@@ -45,6 +46,11 @@ export const routes: Routes = [
         canActivate: [adminGuard],
         path: ProjectLinks.URL_ADMIN_DASHBOARD,
         component: AdminDashboardComponent,
+    },
+    {
+        path: ProjectLinks.URL_QUERY_EXPLAINER,
+        component: QueryExplainerComponent,
+        loadChildren: () => import("./query-explainer/query-explainer.module").then((m) => m.QueryExplainerModule),
     },
     {
         path:
@@ -118,7 +124,7 @@ export const routes: Routes = [
 ];
 
 @NgModule({
-    imports: [RouterModule.forRoot(routes)],
+    imports: [RouterModule.forRoot(routes, { onSameUrlNavigation: "reload" })],
     exports: [RouterModule],
 })
 export class AppRoutingModule {}