+++ src/app/common/app.helpers.ts
@@ -57,7 +57,11 @@ export function momentConvertDateToLocalWithFormat(dateParams: {
 }
 
 export function parseCurrentSchema(data: MaybeNullOrUndefined<DataSchema>): MaybeNull<DatasetSchema> {
-    return data ? (JSON.parse(data.content) as DatasetSchema) : null;
+    return data ? (JSON.parse(removeAllLineBreaks(data.content)) as DatasetSchema) : null;
+}
+
+export function removeAllLineBreaks(value: string): string {
+    return value.replace(/(\r\n|\n|\r)/gm, "");
 }
 
 export function cronExpressionNextTime(cronExpression: string): string {
@@ -85,4 +89,21 @@ export function convertSecondsToHumanReadableFormat(seconds: number): string {
 
 export function capitalizeString(value: string): string {
     return value[0].toUpperCase() + value.slice(1).toLowerCase();
+}
+
+// After the set time in the timeout, the class("clipboard-btn--success") is removed.
+// The result will be a change in the icon. The function looks like a swap
+export function changeCopyIcon(event: MouseEvent): void {
+    if (event.currentTarget !== null) {
+        const htmlButtonElement: HTMLButtonElement = event.currentTarget as HTMLButtonElement;
+        const iconsPair: HTMLCollectionOf<HTMLElement> = htmlButtonElement.children as HTMLCollectionOf<HTMLElement>;
+        setTimeout(() => {
+            iconsPair[0].style.display = "inline-block";
+            iconsPair[1].style.display = "none";
+            htmlButtonElement.classList.remove("clipboard-btn--success");
+        }, AppValues.LONG_DELAY_MS);
+        iconsPair[0].style.display = "none";
+        iconsPair[1].style.display = "inline-block";
+        htmlButtonElement.classList.add("clipboard-btn--success");
+    }
 }