+++ src/app/components/data-access-panel/data-access-panel.component.ts
@@ -2,9 +2,9 @@ import { MaybeUndefined } from "./../../common/app.types";
 import { ProtocolsService } from "./../../services/protocols.service";
 import { ChangeDetectionStrategy, Component, inject, Input, OnInit } from "@angular/core";
 import { DatasetBasicsFragment, DatasetEndpoints } from "src/app/api/kamu.graphql.interface";
-import AppValues from "src/app/common/app.values";
 import { Clipboard } from "@angular/cdk/clipboard";
 import { Observable } from "rxjs";
+import { changeCopyIcon } from "src/app/common/app.helpers";
 
 @Component({
     selector: "app-data-access-panel",
@@ -25,19 +25,7 @@ export class DataAccessPanelComponent implements OnInit {
 
     public copyToClipboard(event: MouseEvent, text: string): void {
         this.clipboard.copy(text);
-        if (event.currentTarget !== null) {
-            const currentElement: HTMLButtonElement = event.currentTarget as HTMLButtonElement;
-            const currentElementChildren: HTMLCollectionOf<HTMLElement> =
-                currentElement.children as HTMLCollectionOf<HTMLElement>;
-            setTimeout(() => {
-                currentElementChildren[0].style.display = "inline-block";
-                currentElementChildren[1].style.display = "none";
-                currentElement.classList.remove("clipboard-btn--success");
-            }, AppValues.LONG_DELAY_MS);
-            currentElementChildren[0].style.display = "none";
-            currentElementChildren[1].style.display = "inline-block";
-            currentElement.classList.add("clipboard-btn--success");
-        }
+        changeCopyIcon(event);
     }
 
     private initClipboardHints(): void {