+++ src/app/dataset-view/dataset.component.ts
@@ -27,6 +27,7 @@ export class DatasetComponent extends BaseDatasetDataComponent implements OnInit
     public datasetViewType: DatasetViewTypeEnum = DatasetViewTypeEnum.Overview;
     public readonly DatasetViewTypeEnum = DatasetViewTypeEnum;
     public sqlLoading: boolean = false;
+    public currentHeadBlockHash: string = "";
 
     private mainDatasetQueryComplete$: Subject<DatasetInfo> = new ReplaySubject<DatasetInfo>(1 /* bufferSize */);
 
@@ -63,9 +64,31 @@ export class DatasetComponent extends BaseDatasetDataComponent implements OnInit
             _.isNil(this.datasetBasics) ||
             this.datasetBasics.name !== urlDatasetInfo.datasetName ||
             this.datasetBasics.owner.accountName !== urlDatasetInfo.accountName ||
-            this.datasetViewType === DatasetViewTypeEnum.Flows
+            [
+                DatasetViewTypeEnum.Overview,
+                DatasetViewTypeEnum.Metadata,
+                DatasetViewTypeEnum.Data,
+                DatasetViewTypeEnum.Lineage,
+            ].includes(this.datasetViewType)
         ) {
-            this.requestMainData(urlDatasetInfo);
+            if (this.datasetBasics) {
+                this.datasetService
+                    .isHeadHashBlockChanged(this.datasetBasics)
+                    .pipe(
+                        switchMap((isNewHead: boolean) => {
+                            return isNewHead ||
+                                this.datasetBasics?.name !== urlDatasetInfo.datasetName ||
+                                this.datasetBasics.owner.accountName !== urlDatasetInfo.accountName
+                                ? this.datasetService.requestDatasetMainData(urlDatasetInfo)
+                                : of();
+                        }),
+                        tap(() => {
+                            this.mainDatasetQueryComplete$.next(urlDatasetInfo);
+                        }),
+                        takeUntilDestroyed(this.destroyRef),
+                    )
+                    .subscribe();
+            }
         }
     }
 
@@ -143,6 +166,7 @@ export class DatasetComponent extends BaseDatasetDataComponent implements OnInit
                 ),
                 first(),
                 switchMap((info: DatasetInfo) => {
+                    /* istanbul ignore else */
                     if (this.datasetViewType === DatasetViewTypeEnum.History) {
                         return this.datasetService.requestDatasetHistory(info, 20, currentPage - 1);
                     } else {
@@ -165,6 +189,7 @@ export class DatasetComponent extends BaseDatasetDataComponent implements OnInit
                 ),
                 first(),
                 switchMap((info) => {
+                    /* istanbul ignore else */
                     if (this.datasetViewType === DatasetViewTypeEnum.Lineage) {
                         return this.datasetService.requestDatasetLineage(info);
                     } else {
@@ -218,6 +243,7 @@ export class DatasetComponent extends BaseDatasetDataComponent implements OnInit
 
     public onClickLineageNode(node: Node): void {
         const nodeData: LineageGraphNodeData = node.data as LineageGraphNodeData;
+        /* istanbul ignore else */
         if (nodeData.kind === LineageGraphNodeKind.Dataset) {
             this.onSelectDataset(nodeData.dataObject.accountName, nodeData.dataObject.name);
         } else {